HHH-20775 Alternative version of https://github.com/hibernate/hibernate-orm/pull/13363 that doesn't use marker files but just checks if there is a manually written MetaModel file
Replace private sourceFile() copies with TestUtil.getSourceFile()

+++ tooling/metamodel-generator/src/test/java/org/hibernate/processor/test/util/TestUtil.java
@@ -269,6 +269,93 @@ public class TestUtil {
 		return outBaseDir;
 	}
 
+	public static File getSourceFile(Class<?> type) {
+		return getSourceFile( type, getSourceBaseDir( type ) );
+	}
+
+	/**
+	 * Returns the source path for a type under the given directory, whether or not the file exists.
+	 */
+	public static File getSourceFile(Class<?> type, File sourceDir) {
+		return new File(
+				sourceDir,
+				type.getName().replace( '.', File.separatorChar ) + ".java"
+		);
+	}
+
+	public static File getClassFile(Class<?> type, File classesDir) {
+		return getClassFile( type.getName(), classesDir );
+	}
+
+	private static File getClassFile(String binaryName, File classesDir) {
+		return new File( classesDir, binaryName.replace( '.', File.separatorChar ) + ".class" );
+	}
+
+	public static File getMetaModelClassFileFor(Class<?> entity, boolean prefix, File classesDir) {
+		return getClassFile( getMetaModelClassName( entity, prefix ), classesDir );
+	}
+
+	public static File getNestedRepositoryClassFileFor(Class<?> repository, boolean prefix, File classesDir) {
+		return getClassFile( getMetaModelClassName( repository.getEnclosingClass(), prefix )
+				+ "$_" + repository.getSimpleName(), classesDir );
+	}
+
+	/**
+	 * Compile the given source files with {@link org.hibernate.processor.HibernateProcessor},
+	 * writing output to {@code outDir} and placing {@code outDir} on the classpath so that
+	 * previously generated classes are visible (as happens in incremental Maven builds).
+	 */
+	public static void compile(File outDir, File... sourceFiles) throws Exception {
+		compile( outDir, outDir, sourceFiles );
+	}
+
+	/**
+	 * Compile with separate class and generated-source output directories, keeping
+	 * previously generated classes on the classpath for incremental builds.
+	 */
+	public static void compile(File classesDir, File generatedSourcesDir, File[] sourceFiles) throws Exception {
+		compile( classesDir, generatedSourcesDir, null,
+				List.of( "-processor", org.hibernate.processor.HibernateProcessor.class.getName() ), sourceFiles );
+	}
+
+	public static void compile(File classesDir, File generatedSourcesDir, File sourceDir, File[] sourceFiles)
+			throws Exception {
+		compile( classesDir, generatedSourcesDir, sourceDir,
+				List.of( "-processor", org.hibernate.processor.HibernateProcessor.class.getName() ), sourceFiles );
+	}
+
+	public static void compileWithoutProcessor(File outDir, File... sourceFiles) throws Exception {
+		compile( outDir, outDir, null, List.of( "-proc:none" ), sourceFiles );
+	}
+
+	private static void compile(File classesDir, File generatedSourcesDir, File sourceDir,
+			List<String> options, File[] sourceFiles)
+			throws Exception {
+		var compiler = javax.tools.ToolProvider.getSystemJavaCompiler();
+		var diagnostics = new javax.tools.DiagnosticCollector<javax.tools.JavaFileObject>();
+		try ( var fileManager = compiler.getStandardFileManager( diagnostics, null, null ) ) {
+			fileManager.setLocation( javax.tools.StandardLocation.CLASS_OUTPUT, List.of( classesDir ) );
+			fileManager.setLocation( javax.tools.StandardLocation.SOURCE_OUTPUT, List.of( generatedSourcesDir ) );
+			if ( sourceDir != null ) {
+				fileManager.setLocation( javax.tools.StandardLocation.SOURCE_PATH, List.of( sourceDir ) );
+			}
+			var classpath = new java.util.ArrayList<File>();
+			for ( File f : fileManager.getLocation( javax.tools.StandardLocation.CLASS_PATH ) ) {
+				classpath.add( f );
+			}
+			classpath.add( classesDir );
+			fileManager.setLocation( javax.tools.StandardLocation.CLASS_PATH, classpath );
+
+			var task = compiler.getTask(
+					null, fileManager, diagnostics,
+					options,
+					null,
+					fileManager.getJavaFileObjectsFromFiles( List.of( sourceFiles ) )
+			);
+			assertTrue( task.call(), "Compilation failed: " + diagnostics.getDiagnostics() );
+		}
+	}
+
 	public static File getSourceBaseDir(Class<?> testClass) {
 		return getBaseDir( testClass, "java" );
 	}
@@ -350,14 +437,18 @@ public class TestUtil {
 	}
 
 	public static File getMetaModelSourceFileFor(Class<?> clazz, boolean prefix) {
+		return getMetaModelSourceFileFor( clazz, prefix, getOutBaseDir( clazz ) );
+	}
+
+	public static File getMetaModelSourceFileFor(Class<?> clazz, boolean prefix, File sourceDir) {
 		if ( clazz.isMemberClass() ) {
-			return getMetaModelSourceFileFor( clazz.getEnclosingClass(), prefix );
+			return getMetaModelSourceFileFor( clazz.getEnclosingClass(), prefix, sourceDir );
 		}
 		String metaModelClassName = getMetaModelClassName(clazz, prefix);
 		// generate the file name
 		String fileName = metaModelClassName.replace( PACKAGE_SEPARATOR, PATH_SEPARATOR );
 		fileName = fileName.concat( ".java" );
-		return new File( getOutBaseDir( clazz ), fileName );
+		return new File( sourceDir, fileName );
 	}
 
 	public static File getMetaModelSourceFileFor(String className) {
@@ -559,7 +650,11 @@ public class TestUtil {
 		}
 	}
 
-	private static void deleteFilesRecursive(File file) {
+	public static void deleteFilesRecursive(File file) {
+		if ( !file.exists() ) {
+			return;
+		}
+		// Delete children before their parent: a directory must be empty before deletion.
 		if ( file.isDirectory() ) {
 			for ( File c : file.listFiles() ) {
 				deleteFilesRecursive( c );
